fix yellow
git-svn-id: svn://svn.rockbox.org/rockbox/trunk@24962 a1c6a512-1295-4272-9138-f99709370657

+++ apps/codecs/au.c
@@ -222,7 +222,7 @@ next_track:
     codec = get_au_codec(format.formattag);
     if (!codec)
     {
-        DEBUGF("CODEC_ERROR: unsupport sun audio format: %x\n", format.formattag);
+        DEBUGF("CODEC_ERROR: unsupport sun audio format: %x\n", (int)format.formattag);
         status = CODEC_ERROR;
         goto done;
     }
